+++ src/main/java/org/datanucleus/store/types/wrappers/LinkedList.java
@@ -22,6 +22,10 @@ import java.util.Collection;
 import java.util.Iterator;
 import java.util.ListIterator;
 
+import org.datanucleus.flush.CollectionAddOperation;
+import org.datanucleus.flush.CollectionRemoveOperation;
+import org.datanucleus.flush.ListAddAtOperation;
+import org.datanucleus.flush.ListRemoveAtOperation;
 import org.datanucleus.metadata.AbstractMemberMetaData;
 import org.datanucleus.state.FetchPlanState;
 import org.datanucleus.state.ObjectProvider;
@@ -388,6 +392,10 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
             ownerOP.getExecutionContext().getRelationshipManager(ownerOP).relationAdd(ownerMmd.getAbsoluteFieldNumber(), element);
         }
 
+        if (SCOUtils.useQueuedUpdate(ownerOP))
+        {
+            ownerOP.getExecutionContext().addOperationToQueue(new ListAddAtOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), index, element));
+        }
         makeDirty();
         if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
         {
@@ -410,6 +418,10 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
 
         if (success)
         {
+            if (SCOUtils.useQueuedUpdate(ownerOP))
+            {
+                ownerOP.getExecutionContext().addOperationToQueue(new CollectionAddOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element));
+            }
             makeDirty();
             if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
             {
@@ -440,6 +452,13 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
 
         if (success)
         {
+            if (SCOUtils.useQueuedUpdate(ownerOP))
+            {
+                for (Object element : elements)
+                {
+                    ownerOP.getExecutionContext().addOperationToQueue(new CollectionAddOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element));
+                }
+            }
             makeDirty();
             if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
             {
@@ -471,6 +490,14 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
 
         if (success)
         {
+            if (SCOUtils.useQueuedUpdate(ownerOP))
+            {
+                int pos = index;
+                for (Object element : elements)
+                {
+                    ownerOP.getExecutionContext().addOperationToQueue(new ListAddAtOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), pos++, element));
+                }
+            }
             makeDirty();
             if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
             {
@@ -501,6 +528,10 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
     public void addLast(Object element)
     {
         delegate.addLast(element);
+        if (SCOUtils.useQueuedUpdate(ownerOP))
+        {
+            ownerOP.getExecutionContext().addOperationToQueue(new CollectionAddOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element));
+        }
         makeDirty();
         if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
         {
@@ -533,7 +564,15 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
                 Iterator iter = copy.iterator();
                 while (iter.hasNext())
                 {
-                    ownerOP.getExecutionContext().deleteObjectInternal(iter.next());
+                    Object element = iter.next();
+                    if (SCOUtils.useQueuedUpdate(ownerOP))
+                    {
+                        ownerOP.getExecutionContext().addOperationToQueue(new CollectionRemoveOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element, true));
+                    }
+                    else
+                    {
+                        ownerOP.getExecutionContext().deleteObjectInternal(element);
+                    }
                 }
             }
         }
@@ -563,7 +602,14 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
         if (ownerOP != null && SCOUtils.hasDependentElement(ownerMmd))
         {
             // Cascade delete
-            ownerOP.getExecutionContext().deleteObjectInternal(element);
+            if (SCOUtils.useQueuedUpdate(ownerOP))
+            {
+                ownerOP.getExecutionContext().addOperationToQueue(new CollectionRemoveOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element, true));
+            }
+            else
+            {
+                ownerOP.getExecutionContext().deleteObjectInternal(element);
+            }
         }
 
         makeDirty();
@@ -601,7 +647,14 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
             SCOUtils.hasDependentElement(ownerMmd))
         {
             // Cascade delete
-            ownerOP.getExecutionContext().deleteObjectInternal(element);
+            if (SCOUtils.useQueuedUpdate(ownerOP))
+            {
+                ownerOP.getExecutionContext().addOperationToQueue(new CollectionRemoveOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element, allowCascadeDelete));
+            }
+            else
+            {
+                ownerOP.getExecutionContext().deleteObjectInternal(element);
+            }
         }
 
         if (success)
@@ -643,7 +696,15 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
                 Iterator iter = elements.iterator();
                 while (iter.hasNext())
                 {
-                    ownerOP.getExecutionContext().deleteObjectInternal(iter.next());
+                    Object element = iter.next();
+                    if (SCOUtils.useQueuedUpdate(ownerOP))
+                    {
+                        ownerOP.getExecutionContext().addOperationToQueue(new CollectionRemoveOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element, true));
+                    }
+                    else
+                    {
+                        ownerOP.getExecutionContext().deleteObjectInternal(element);
+                    }
                 }
             }
         }
@@ -674,6 +735,10 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
             ownerOP.getExecutionContext().deleteObjectInternal(element);
         }
 
+        if (SCOUtils.useQueuedUpdate(ownerOP))
+        {
+            ownerOP.getExecutionContext().addOperationToQueue(new ListRemoveAtOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), 0));
+        }
         makeDirty();
         if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
         {
@@ -696,6 +761,10 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
             ownerOP.getExecutionContext().deleteObjectInternal(element);
         }
 
+        if (SCOUtils.useQueuedUpdate(ownerOP))
+        {
+            ownerOP.getExecutionContext().addOperationToQueue(new ListRemoveAtOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), 0));
+        }
         makeDirty();
         if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
         {
@@ -734,13 +803,20 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
      */
     public Object set(int index, Object element, boolean allowDependentField)
     {
-        Object obj = delegate.set(index, element);
-        if (ownerOP != null && allowDependentField && !delegate.contains(obj))
+        Object prevElement = delegate.set(index, element);
+        if (ownerOP != null && allowDependentField && !delegate.contains(prevElement))
         {
             // Cascade delete
             if (SCOUtils.hasDependentElement(ownerMmd))
             {
-                ownerOP.getExecutionContext().deleteObjectInternal(obj);
+                if (SCOUtils.useQueuedUpdate(ownerOP))
+                {
+                    ownerOP.getExecutionContext().addOperationToQueue(new ListRemoveAtOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), index));
+                }
+                else
+                {
+                    ownerOP.getExecutionContext().deleteObjectInternal(prevElement);
+                }
             }
         }
 
@@ -749,7 +825,7 @@ public class LinkedList extends java.util.LinkedList implements SCOList<java.uti
         {
             ownerOP.getExecutionContext().processNontransactionalUpdate();
         }
-        return obj;
+        return prevElement;
     }
 
     /**
