+++ src/main/java/org/datanucleus/store/types/wrappers/List.java
@@ -23,6 +23,10 @@ import java.util.Collection;
 import java.util.Iterator;
 import java.util.ListIterator;
 
+import org.datanucleus.flush.CollectionAddOperation;
+import org.datanucleus.flush.CollectionRemoveOperation;
+import org.datanucleus.flush.ListAddAtOperation;
+import org.datanucleus.flush.ListRemoveAtOperation;
 import org.datanucleus.metadata.AbstractMemberMetaData;
 import org.datanucleus.state.FetchPlanState;
 import org.datanucleus.state.ObjectProvider;
@@ -382,6 +386,10 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
 
         if (success)
         {
+            if (SCOUtils.useQueuedUpdate(ownerOP))
+            {
+                ownerOP.getExecutionContext().addOperationToQueue(new CollectionAddOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element));
+            }
             makeDirty();
             if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
             {
@@ -404,6 +412,10 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
             ownerOP.getExecutionContext().getRelationshipManager(ownerOP).relationAdd(ownerMmd.getAbsoluteFieldNumber(), element);
         }
 
+        if (SCOUtils.useQueuedUpdate(ownerOP))
+        {
+            ownerOP.getExecutionContext().addOperationToQueue(new ListAddAtOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), index, element));
+        }
         makeDirty();
         if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
         {
@@ -432,6 +444,13 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
 
         if (success)
         {
+            if (SCOUtils.useQueuedUpdate(ownerOP))
+            {
+                for (Object element : elements)
+                {
+                    ownerOP.getExecutionContext().addOperationToQueue(new CollectionAddOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element));
+                }
+            }
             makeDirty();
             if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
             {
@@ -463,6 +482,14 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
 
         if (success)
         {
+            if (SCOUtils.useQueuedUpdate(ownerOP))
+            {
+                int pos = index;
+                for (Object element : elements)
+                {
+                    ownerOP.getExecutionContext().addOperationToQueue(new ListAddAtOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), pos++, element));
+                }
+            }
             makeDirty();
             if (ownerOP != null && !ownerOP.getExecutionContext().getTransaction().isActive())
             {
@@ -497,7 +524,15 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
                 Iterator iter = copy.iterator();
                 while (iter.hasNext())
                 {
-                    ownerOP.getExecutionContext().deleteObjectInternal(iter.next());
+                    Object element = iter.next();
+                    if (SCOUtils.useQueuedUpdate(ownerOP))
+                    {
+                        ownerOP.getExecutionContext().addOperationToQueue(new CollectionRemoveOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element, true));
+                    }
+                    else
+                    {
+                        ownerOP.getExecutionContext().deleteObjectInternal(element);
+                    }
                 }
             }
         }
@@ -539,7 +574,14 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
             // Cascade delete
             if (SCOUtils.hasDependentElement(ownerMmd))
             {
-                ownerOP.getExecutionContext().deleteObjectInternal(element);
+                if (SCOUtils.useQueuedUpdate(ownerOP))
+                {
+                    ownerOP.getExecutionContext().addOperationToQueue(new CollectionRemoveOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element, allowCascadeDelete));
+                }
+                else
+                {
+                    ownerOP.getExecutionContext().deleteObjectInternal(element);
+                }
             }
         }
 
@@ -573,7 +615,14 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
             // Cascade delete
             if (SCOUtils.hasDependentElement(ownerMmd))
             {
-                ownerOP.getExecutionContext().deleteObjectInternal(element);
+                if (SCOUtils.useQueuedUpdate(ownerOP))
+                {
+                    ownerOP.getExecutionContext().addOperationToQueue(new ListRemoveAtOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), index));
+                }
+                else
+                {
+                    ownerOP.getExecutionContext().deleteObjectInternal(element);
+                }
             }
         }
 
@@ -612,7 +661,15 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
                 Iterator iter = elements.iterator();
                 while (iter.hasNext())
                 {
-                    ownerOP.getExecutionContext().deleteObjectInternal(iter.next());
+                    Object element = iter.next();
+                    if (SCOUtils.useQueuedUpdate(ownerOP))
+                    {
+                        ownerOP.getExecutionContext().addOperationToQueue(new CollectionRemoveOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), element, true));
+                    }
+                    else
+                    {
+                        ownerOP.getExecutionContext().deleteObjectInternal(element);
+                    }
                 }
             }
         }
@@ -659,13 +716,20 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
      */
     public Object set(int index, Object element, boolean allowDependentField)
     {
-        Object obj = delegate.set(index, element);
-        if (ownerOP != null && allowDependentField && !delegate.contains(obj))
+        Object prevElement = delegate.set(index, element);
+        if (ownerOP != null && allowDependentField && !delegate.contains(prevElement))
         {
             // Cascade delete
             if (SCOUtils.hasDependentElement(ownerMmd))
             {
-                ownerOP.getExecutionContext().deleteObjectInternal(obj);
+                if (SCOUtils.useQueuedUpdate(ownerOP))
+                {
+                    ownerOP.getExecutionContext().addOperationToQueue(new ListRemoveAtOperation(ownerOP, ownerMmd.getAbsoluteFieldNumber(), index));
+                }
+                else
+                {
+                    ownerOP.getExecutionContext().deleteObjectInternal(prevElement);
+                }
             }
         }
 
@@ -674,7 +738,7 @@ public class List extends AbstractList implements SCOList<java.util.List>, Clone
         {
             ownerOP.getExecutionContext().processNontransactionalUpdate();
         }
-        return obj;
+        return prevElement;
     }
 
     /**
